requirements.txt
streamlit>=1.42.0
pandas
gspread
google-auth
Authlib>=1.3.2